+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#Unreleased
+- [...]
+
 #v0.1.0
 - [Transitions] Add Transitions component
 - [Modal] Apply Transitions component and add unit tests